Added shuffleboard variables
