+++ Custom_Swerve/src/main/java/frc/robot/Robot.java
@@ -4,6 +4,7 @@
 
 package frc.robot;
 
+import edu.wpi.first.wpilibj.CAN;
 import edu.wpi.first.wpilibj.TimedRobot;
 import edu.wpi.first.wpilibj.XboxController;
 
@@ -21,11 +22,13 @@ public class Robot extends TimedRobot {
    * This function is run when the robot is first started up and should be used for any
    * initialization code.
    */
+  CANcoder coder1 = new CANcoder(1);
+  CANcoder coder3 = new CANcoder(3);
   SwerveModule moduleFL = new SwerveModule(new TalonFX(10),new TalonFX(11),new CANcoder(1),0);
   SwerveModule moduleFR = new SwerveModule(new TalonFX(40),new TalonFX(41),new CANcoder(4),0);
-  SwerveModule moduleBL = new SwerveModule(new TalonFX(20),new TalonFX(21),new CANcoder(2),0);
-  SwerveModule moduleBR = new SwerveModule(new TalonFX(30),new TalonFX(31),new CANcoder(3),0);
-  Drivetrain robot = new Drivetrain(moduleFL,moduleFR,moduleBL,moduleBR);
+  SwerveModule moduleBL = new SwerveModule(new TalonFX(20),new TalonFX(21),new CANcoder(2),0.25);
+  SwerveModule moduleBR = new SwerveModule(new TalonFX(30),new TalonFX(31),new CANcoder(3),0.25);
+  DriveTrain robot = new DriveTrain(moduleFL,moduleFR,moduleBL,moduleBR);
   XboxController joysticks = new XboxController(0);
   @Override
   public void robotInit(){
@@ -50,8 +53,10 @@ public class Robot extends TimedRobot {
   @Override
   public void teleopPeriodic() {
       //moduleFL.TurnTo(controller.getRawAxis(1));
-      robot.setDriveDirection(joysticks.getRawAxis(2));
-      robot.setSpeed(joysticks.getRawAxis(1)/2);
+      robot.setDriveDirection(joysticks.getRawAxis(0));
+      robot.setSpeed(joysticks.getRawAxis(1)/5);
+      double coder1reading = coder1.getPosition().getValueAsDouble();
+      double coder3reading = coder3.getPosition().getValueAsDouble();
   }
 
   @Override
